+++ http.gemspec
@@ -30,8 +30,7 @@ Gem::Specification.new do |gem|
   gem.add_runtime_dependency "addressable",    "~> 2.3"
   gem.add_runtime_dependency "http-cookie",    "~> 1.0"
   gem.add_runtime_dependency "http-form_data", "~> 2.2"
-  gem.add_runtime_dependency "http-parser",    "~> 1.2.0"
-  gem.add_runtime_dependency "llhttp"
+  gem.add_runtime_dependency "llhttp-ffi",     "~> 0.0.1"
 
   gem.add_development_dependency "bundler", "~> 2.0"
